initial setup, add class for player and rooms, creating method for player moving around rooms
