+++ src/old_src/item.py
@@ -1,9 +1,7 @@
-class Item: 
+
+class Item:
     def __init__(self, name, description):
         self.name = name
         self.description = description
-
     def __str__(self):
-        return f"Item: {self.name}, Description: {self.description}"
-
-    
+        return "Name of the item: {self.name}, Description of the item: {self.description}".format(self=self)
